+++ packages/core/src/product/mod.ts
@@ -1,0 +1,18 @@
+export {
+  USER_STORIES,
+  pickStory,
+  type StoryId,
+  type UserStory,
+  type SituationSnapshot,
+} from "./stories.js";
+export {
+  detectSituation,
+  type KitSituation,
+  type DetectSituationOptions,
+} from "./situation.js";
+export {
+  runReady,
+  type ReadyOptions,
+  type ReadyReport,
+  type ReadyResult,
+} from "./ready.js";